+++ src/xrGame/WeaponAutomaticShotgun.h
@@ -31,13 +31,13 @@ public:
 	void			PlayAnimCloseWeapon		();
 
 	virtual	int		GetCurrentFireMode	() { return m_aFireModes[m_iCurFireMode]; };
-	virtual void	OnMotionMark(u32 state, const motion_marks&);
+	virtual void	OnMotionMark(u8 state, const motion_marks&);
 
 
 protected:
-	virtual void	OnAnimationEnd			(u32 state);
+	virtual void	OnAnimationEnd			(u8 state);
 	void			TriStateReload			();
-	virtual void	OnStateSwitch			(u32 S);
+	virtual void	OnStateSwitch			(u8 S);
 
 	ESoundTypes		m_eSoundOpen;
 	ESoundTypes		m_eSoundAddCartridge;
