+++ src/xrGame/hud_item_object.h
@@ -52,8 +52,8 @@ public:
 public:
 	virtual void		Load				(const char* section);
 	virtual bool		Action				(u16 cmd, u32 flags);
-	virtual void		SwitchState			(u32 S);
-	virtual void		OnStateSwitch		(u32 S);
+	virtual void		SwitchState			(u8 S);
+	virtual void		OnStateSwitch		(u8 S);
 	virtual void		OnEvent				(NET_Packet& P, u16 type);
 	virtual void		OnH_A_Chield		();
 	virtual void		OnH_B_Chield		();
